Move new interfaces into sugar package, move sugar methods to bottom of Spark
